Compute weighted strength reps

+++ app_workout/models.py
@@ -266,7 +266,7 @@ class StrengthDailyLog(models.Model):
     datetime_started = models.DateTimeField()
     routine = models.ForeignKey(StrengthRoutine, on_delete=models.PROTECT, related_name="daily_logs")
     rep_goal = models.FloatField(null=True, blank=True)
-    total_reps_completed = models.PositiveIntegerField(null=True, blank=True)
+    total_reps_completed = models.FloatField(null=True, blank=True)
     max_reps = models.PositiveIntegerField(null=True, blank=True)
     max_weight = models.FloatField(null=True, blank=True)
     minutes_elapsed = models.FloatField(null=True, blank=True)
